backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pillow==10.3.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
ultralytics==8.3.0
numpy==1.26.4
